Fix -Wsign-compare

+++ engine/src/render/fancyterraindata.cpp
@@ -360,7 +360,7 @@ std::tuple<Vector3f, bool> isect_terrain_ray(
         int x, y;
         std::tie(x, y) = *iter;
 
-        if (x < 0 || y < 0 || x >= size-1 || y >= size-1) {
+        if (x < 0 || y < 0 || x >= (int)(size-1) || y >= (int)(size-1)) {
             continue;
         }
